refactor: WebArchiveJob 클래스를 도입한다
 - 새로운 블로그의 인덱스 페이지를 만들기 위해 새로운 BulkRequestSender,
   ArchivePageService, IndexPageGenerator 클래스를 정의해야 한다
 - WebArchiveScheduler에서 모든 인덱스 페이지를 만들 수 있게 하려면,
   기존 클래스와 새로 만든 클래스를 모두 알고 있어야 한다.  이렇게
   바꿨을 경우, doSend() 등 WebArchiveScheduler 내부에서 호출하는 메서드에
   속성이 추가되어야 해서 유지보수가 어려워진다
 - 각 블로그 인덱스 페이지 생성을 담당하는 빈을 갖는 WebArchiveJob을 정의해서
   WebArchiveScheduler는 단순히 WebArchiveJob에 요청만 하도록 한다
